requirements.txt
argcomplete==3.6.2
blinker==1.9.0
boto3==1.38.41
botocore==1.38.41
cachetools==5.5.2
certifi==2025.6.15
cfn-flip==1.3.0
charset-normalizer==3.4.2
click==8.2.1
durationpy==0.10
Flask==3.1.1
flask-cors==6.0.1
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
hjson==3.1.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==3.0.2
oauthlib==3.3.1
placebo==0.9.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
troposphere==4.9.3
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
zappa==0.60.2
